[forum creation page] moving to separate component
